requirements.txt
# Core dependencies
discord.py>=2.3.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0

# Database
motor>=3.3.0
pymongo>=4.5.0

# Cache
redis>=5.0.0

# HTTP and networking
aiohttp>=3.8.0
requests>=2.31.0

# Image processing
Pillow>=10.0.0

# AI and external services
openai>=1.0.0
anthropic>=0.7.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Production
gunicorn>=21.0.0 